renamed kuzzleCollection.js -> Collection.js
