+++ src/Collection.js
